+++ src/api/controller/admin_users_controller.rs
@@ -7,8 +7,8 @@ use crate::domain::crypto::SchemeAwareHasher;
 use crate::domain::error::UserError;
 use crate::domain::event::UserEvents;
 use crate::domain::jwt::UserDTO;
-use crate::domain::repository::RepositoryError;
 use crate::domain::user::{PasswordHandler, User};
+use crate::infrastructure::repository::RepositoryError;
 use axum::Json;
 use axum::extract::{Path, Query, State};
 use axum::http::StatusCode;
