+++ src/bin/app.rs
@@ -7,13 +7,13 @@ use auth_service::application::service::auth_service::AuthService;
 use auth_service::domain::crypto::SchemeAwareHasher;
 use auth_service::domain::error::UserError;
 use auth_service::domain::event::UserEvents;
-use auth_service::domain::repository::RepositoryError;
 use auth_service::domain::role::Role;
 use auth_service::domain::user::{PasswordHandler, User};
 use auth_service::infrastructure::database::create_pool;
 use auth_service::infrastructure::message_consumer::MessageConsumer;
 use auth_service::infrastructure::message_publisher::MessagePublisher;
 use auth_service::infrastructure::permission_repository::PermissionRepository;
+use auth_service::infrastructure::repository::RepositoryError;
 use auth_service::infrastructure::role_repository::RoleRepository;
 use auth_service::infrastructure::user_repository::UserRepository;
 use clap::{Parser, Subcommand};
