+++ src/domain/mod.rs
@@ -3,6 +3,5 @@ pub mod error;
 pub mod event;
 pub mod jwt;
 pub mod permission;
-pub mod repository;
 pub mod role;
 pub mod user;
