--- src/domain/repository.rs
@@ -1,31 +1,0 @@
-use sqlx::Error as SqlxError;
-use std::error::Error;
-use std::fmt;
-
-#[derive(Debug)]
-pub enum RepositoryError {
-    NotFound(String),
-    Database(SqlxError),
-    Conflict(String),
-    ValidationError(String),
-}
-
-impl fmt::Display for RepositoryError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            RepositoryError::NotFound(msg) => write!(f, "{}", msg),
-            RepositoryError::Database(e) => write!(f, "Database error: {}", e),
-            RepositoryError::Conflict(msg) => write!(f, "Conflict error: {}", msg),
-            RepositoryError::ValidationError(msg) => write!(f, "Validation error: {}", msg),
-        }
-    }
-}
-
-impl Error for RepositoryError {
-    fn source(&self) -> Option<&(dyn Error + 'static)> {
-        match self {
-            RepositoryError::Database(e) => Some(e),
-            _ => None,
-        }
-    }
-}
