+++ src/infrastructure/sqlite_role_repository.rs
@@ -1,7 +1,7 @@
 use crate::domain::permission::Permission;
-use crate::domain::repository::RepositoryError;
 use crate::domain::role::Role;
 use crate::infrastructure::dto::RoleWithPermissionsRow;
+use crate::infrastructure::repository::RepositoryError;
 use sqlx::{Pool, Sqlite, query_as};
 use std::collections::HashMap;
 use uuid::Uuid;
