+++ src/infrastructure/sqlite_user_repository.rs
@@ -1,8 +1,8 @@
 use crate::domain::permission::Permission;
-use crate::domain::repository::RepositoryError;
 use crate::domain::role::Role;
 use crate::domain::user::User;
 use crate::infrastructure::dto::{UserWithPermissionsRow, UserWithRoleRow};
+use crate::infrastructure::repository::RepositoryError;
 use sqlx::{Error, Pool, Sqlite, query};
 use std::collections::{HashMap, HashSet};
 use uuid::Uuid;
