+++ tests/integration/permission_repository_test.rs
@@ -1,6 +1,6 @@
 use crate::utils::runners::run_database_test_with_default;
 use auth_service::domain::permission::Permission;
-use auth_service::domain::repository::RepositoryError;
+use auth_service::infrastructure::repository::RepositoryError;
 use uuid::Uuid;
 
 #[tokio::test]
