+++ tests/integration/role_repository_test.rs
@@ -1,7 +1,7 @@
 use crate::utils::runners::run_database_test_with_default;
 use auth_service::domain::permission::Permission;
-use auth_service::domain::repository::RepositoryError;
 use auth_service::domain::role::Role;
+use auth_service::infrastructure::repository::RepositoryError;
 use uuid::Uuid;
 
 #[tokio::test]
